レビュー対応: テスト整理とgitignore更新
- test-results/ を .gitignore に追加（生成物をリポジトリから除外）
- vite-tsconfig-paths を削除（未使用）
- makeQueryMock を __tests__/helpers/queryMock.ts に切り出し
- POST テストに user_id が正しくセットされることの検証を追加
