+++ __tests__/api/expense.test.ts
@@ -1,5 +1,6 @@
 import { NextRequest } from 'next/server';
 import { describe, it, expect, vi, beforeEach } from 'vitest';
+import { makeQueryMock } from '../helpers/queryMock';
 
 const mockFrom = vi.fn();
 const mockGetUser = vi.fn();
@@ -20,24 +21,6 @@ vi.mock('next/headers', () => ({
 import { GET, POST, PATCH, DELETE } from '../../app/api/expense/route';
 
 const TEST_USER = { id: 'user-123', email: 'test@example.com' };
-
-function makeQueryMock(result: object) {
-  const q: Record<string, unknown> = {};
-  const chain = () => q;
-  q['then'] = (resolve: (v: unknown) => unknown, reject?: (r: unknown) => unknown) =>
-    Promise.resolve(result).then(resolve, reject);
-  q.select = vi.fn(chain);
-  q.insert = vi.fn(chain);
-  q.update = vi.fn(chain);
-  q.delete = vi.fn(chain);
-  q.ilike = vi.fn(chain);
-  q.gte = vi.fn(chain);
-  q.lte = vi.fn(chain);
-  q.order = vi.fn(chain);
-  q.range = vi.fn(() => Promise.resolve(result));
-  q.eq = vi.fn(chain);
-  return q;
-}
 
 beforeEach(() => {
   vi.clearAllMocks();
@@ -118,7 +101,8 @@ describe('POST /api/expense', () => {
 
   it('有効なデータで支出を作成できる', async () => {
     const newRow = { id: 1, source: '食費', amount: 5000, category: '食費', entry_date: '2026-04-01', user_id: TEST_USER.id };
-    mockFrom.mockReturnValue(makeQueryMock({ data: [newRow], error: null }));
+    const q = makeQueryMock({ data: [newRow], error: null });
+    mockFrom.mockReturnValue(q);
 
     const req = new Request('http://localhost/api/expense', {
       method: 'POST',
@@ -127,6 +111,9 @@ describe('POST /api/expense', () => {
     const res = await POST(req);
     expect(res.status).toBe(200);
     expect(await res.json()).toMatchObject({ message: '作成OK' });
+    expect(q.insert).toHaveBeenCalledWith(
+      expect.arrayContaining([expect.objectContaining({ user_id: TEST_USER.id })])
+    );
   });
 
   it('sourceが空の場合はバリデーションエラー400', async () => {
