+++ package.json
@@ -34,7 +34,6 @@
     "supabase": "^2.34.3",
     "tailwindcss": "^3.4.17",
     "typescript": "^5.9.2",
-    "vite-tsconfig-paths": "^6.1.1",
     "vitest": "^2.1.9"
   }
 }
